refactor: unify chooseInvite and DisplayInvite pages into single DisplayInvite component
